+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `ViewComboBox` component handle viewmode change.
 * Added `compas_viewer.objects.BufferGeometry` and `compas_viewer.objects.BufferObject`.
 * Added `clear_scene` action.
+* Added `compas_viewer.components.SceneForm`.
+* Added `callback` for `compas_viewer.components.TreeForm`.
 
 ### Changed